+++ src/main/java/template/Part_fashion.java
@@ -8,9 +8,9 @@ public class Part_fashion {
     public static final List<Short> fashions = new ArrayList<>();
     public static final List<Part_fashion> entrys = new ArrayList<>();
     public short id;
-    public byte[] part;
+    public short[] part;
 
-    public static byte[] get_part(Player p) {
+    public static short[] get_part(Player p) {
         if (p.item.wear[11] != null) {
             for (Part_fashion temp : entrys) {
                 if (temp.id == p.item.wear[11].id) {
@@ -25,6 +25,6 @@ public class Part_fashion {
                 }
             }
         }
-        return new byte[]{-1, -1, -1, -1, -1, -1, -1};
+        return new short[]{-1, -1, -1, -1, -1, -1, -1};
     }
 }
